pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "grandalf>=0.8",
    "huggingface-hub>=1.8.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "transformers>=5.4.0",
]
